refactor(cloud): move bridge package under puffo_agent/cloud/
Namespaces the cloud Bridge as puffo_agent.cloud.bridge to distinguish
it from the ws-local WsLocalBridge (the local relay for an attached
tool). Imports updated; no behaviour change.
